editoast: moved utility functions to helpers.rs
